+++ examples/typescript/streaming/run.ts
@@ -1,16 +1,13 @@
-import { RunEventType } from '@hatchet-dev/typescript-sdk-dev/typescript-sdk/clients/listeners/run-listener/child-listener-client';
 import { streamingTask } from './workflow';
+import { hatchet } from '../hatchet-client';
 
 async function main() {
   // > Consume
   const ref = await streamingTask.runNoWait({});
+  const id = await ref.getWorkflowRunId();
 
-  const stream = await ref.stream();
-
-  for await (const event of stream) {
-    if (event.type === RunEventType.STEP_RUN_EVENT_TYPE_STREAM) {
-      process.stdout.write(event.payload);
-    }
+  for await (const content of hatchet.runs.subscribeToStream(id)) {
+    process.stdout.write(content);
   }
 }
