+++ frontend/docs/pages/home/streaming.mdx
@@ -68,4 +68,4 @@ For example, with FastAPI, you'd do the following:
 
 <Snippet src={snips.python.streaming.fastapi_proxy} block="fastapi_proxy" />
 
-Then, assuming you run the server on port `8000`, running `curl -N -X GET http://localhost:8000/stream` would result in the text streaming back to your console from Hatchet through your FastAPI proxy.
+Then, assuming you run the server on port `8000`, running `curl -N http://localhost:8000/stream` would result in the text streaming back to your console from Hatchet through your FastAPI proxy.
